projects/gemini-urls-context/pyproject.toml
[project]
name = "gemini-urls-context"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.31.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
